[Ajout] test tc133Context.php (+quelques légers correctifs)

+++ test/acceptance/tc154Context.php
@@ -22,20 +22,13 @@ class tc154Context implements Context
   {
     $canDelete = Universe::getUniverse()->getCanDelete();
     $userDao = App_DaoFactory::getFactory()->getUserDao();
-    if (isset($canDelete['user'])) {
-      $user = $userDao->selectUserByEmail(Universe::getUniverse()->getUser()->getEmail());
-      $isAdmin = $user->getIsAdmin();
-      if ($isAdmin == false) {
-        $userDao->deleteUser($user->getId());
-      }
-      unset($canDelete['user']);
-      Universe::getUniverse()->setCanDelete($canDelete);
-    }
     if (isset($canDelete['user2'])) {
       $user2 = $userDao->selectUserByEmail(Universe::getUniverse()->getUser2()->getEmail());
-      $isAdmin2 = $user2->getIsAdmin();
-      if ($isAdmin2 == false) {
-        $userDao->deleteUser($user2->getId());
+      if ($user2 != null) {
+        $isAdmin2 = $user2->getIsAdmin();
+        if ($isAdmin2 == false) {
+          $userDao->deleteUser($user2->getId());
+        }
       }
       unset($canDelete['user2']);
       Universe::getUniverse()->setCanDelete($canDelete);
